# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/posts": {
      "get": {
        "operationId": "PostsController_getAllPosts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostsController_getPostById",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/posts/create": {
      "post": {
        "operationId": "PostsController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/posts/update/{id}": {
      "patch": {
        "operationId": "PostsController_updatePostById",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/posts/delete/{id}": {
      "delete": {
        "operationId": "PostsController_deletePostById",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Authorization API",
    "description": "Swagger specification for api-auth | [swagger.json](swagger.json)",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreatePostDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}